LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/request-end-to-end.https-expected.txt


FAIL Test FetchEvent.request passed to onfetch assert_equals: request.mode expected "navigate" but got "no-cors"
